Stuck
With the pom cleaned up of duplicate dependencies, we're now getting
errors about missing functionality.  That makes sense.  But some of the
functionality comes from jars that add a lot of duplication.  In some of
those cases, exclusion is not possible, the whole jar is included by
other fat jars like iceberg-flink-runtime.  The current error with this
version of the pom and a bunch of hive jars copied into my
stat1004:/home/user/milimetric/flink-1.15.2/lib/ folder, is this:

Caused by: org.apache.hadoop.hive.metastore.api.MetaException: Could not connect to meta store using any of the URIs provided. Most recent failure: org.apache.thrift.transport.TTransportException: GSS initiate failed

This goes away if we include the jars we excluded, but then of course we
get the Parquet incompatibilities and other problems.  So it feels like
the build is stuck and we have to thread the needle between duplicate
dependencies and the functionality we need.  This may not be possible
without going into some of the other jars and rebuilding them (like the
iceberg-flink jars specifically, they seem the most problematic).  One
simple thing to try would be to use the iceberg-flink jar and add the
rest of the dependencies manually instead of relying on the fat jar.

I'm personally going to give up on this for now.  I want something to
work by end of day tomorrow, and I feel like the spark streaming
approach is more tractable.